backend/requirements.txt
fastapi==0.115.12
loguru==0.7.3
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic-settings==2.9.1
pymupdf4llm==0.0.22
python-dotenv==1.1.0
SQLAlchemy==2.0.40
python-multipart==0.0.6
uvicorn==0.34.2
llama-index-core==0.12.35
llama-index-llms-openai==0.3.38
llama-index-embeddings-openai==0.3.1
llama-index-vector-stores-postgres==0.5.2
sentence-transformers==4.1.0
torch==2.7.0+cpu